Added cart service

+++ src/main/java/com/shoesapp/cart/entity/Cart.java
@@ -23,6 +23,6 @@ public class Cart {
     @OneToMany(mappedBy = "cart", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
     private List<CartItem> cartItems = new ArrayList<>();
 
-    private BigDecimal totalPrice;
+    private Double totalPrice;
 
 }
